move NonZeroExitCode to ecs package
