+++ internal/pkg/ecs/errors.go
@@ -12,3 +12,13 @@ type ErrMultipleContainersInTaskDef struct {
 func (e *ErrMultipleContainersInTaskDef) Error() string {
 	return fmt.Sprintf("found more than one container in task definition: %s", e.taskDefIdentifier)
 }
+
+type errExitCode struct {
+	containerName string
+	taskId        string
+	exitCode      int64
+}
+
+func (e *errExitCode) Error() string {
+	return fmt.Sprintf("Container %s in task %s exited with status code %d", e.containerName, e.taskId, e.exitCode)
+}
