+++ internal/pkg/task/config_runner.go
@@ -5,7 +5,6 @@ package task
 
 import (
 	"fmt"
-	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/copilot-cli/internal/pkg/aws/ec2"
 	"github.com/aws/copilot-cli/internal/pkg/aws/ecs"
 )
@@ -35,6 +34,9 @@ type ConfigRunner struct {
 
 	// Must not be nil if using default subnets.
 	VPCGetter VPCGetter
+
+	// Figures non-zero exit code of the task
+	NonZeroExitCodeGetter NonZeroExitCodeGetter
 
 	// Platform configuration
 	OS string
@@ -107,40 +109,9 @@ func (r *ConfigRunner) validateDependencies() error {
 
 // CheckNonZeroExitCode returns the status of the containers part of the given tasks.
 func (r *ConfigRunner) CheckNonZeroExitCode(tasks []*Task) (int, error) {
-	essentialContainers := make(map[string]bool)
-	taskDefName := fmt.Sprintf("copilot-%s", r.GroupName)
-	taskDefinition, err := r.Starter.TaskDefinition(taskDefName)
-	if err != nil {
-		return 0, fmt.Errorf("get task definition %s of a task", taskDefName)
-	}
-
-	for _, container := range taskDefinition.ContainerDefinitions {
-		if aws.BoolValue(container.Essential) {
-			essentialContainers[aws.StringValue(container.Name)] = true
-		}
-	}
-
 	taskARNs := make([]string, len(tasks))
 	for idx, task := range tasks {
 		taskARNs[idx] = task.TaskARN
 	}
-	describedTasks, describeErr := r.Starter.DescribeTasks(r.Cluster, taskARNs)
-	if describeErr != nil {
-		return 0, describeErr
-	}
-
-	for _, describedTask := range describedTasks {
-		for _, container := range describedTask.Containers {
-			if essentialContainers[aws.StringValue(container.Name)] && aws.Int64Value(container.ExitCode) > 0 {
-				taskID, err := ecs.TaskID(aws.StringValue(describedTask.TaskArn))
-				if err != nil {
-					return 0, err
-				}
-				return int(aws.Int64Value(container.ExitCode)), &errExitCode{aws.StringValue(container.Name),
-					taskID,
-					aws.Int64Value(container.ExitCode)}
-			}
-		}
-	}
-	return 0, nil
+	return r.NonZeroExitCodeGetter.NonZeroExitCode(taskARNs, r.GroupName, r.Cluster)
 }
