+++ internal/pkg/task/env_runner_test.go
@@ -346,118 +346,3 @@ func TestEnvRunner_Run(t *testing.T) {
 		})
 	}
 }
-
-func TestEnvRunner_CheckNonZeroExitCode_Run(t *testing.T) {
-	testCases := map[string]struct {
-		count     int
-		groupName string
-		follow    bool
-
-		securityGroups []string
-
-		os   string
-		arch string
-
-		mockStarter       func(m *mocks.MockRunner)
-		mockClusterGetter func(m *mocks.MockClusterGetter)
-		inTasks           []*Task
-
-		wantedError error
-	}{
-
-		"returns the non zero exit code of the essential container": {
-			inTasks: []*Task{
-				{
-					TaskARN:    "mockTask1",
-					ClusterARN: "cluster-1",
-				},
-			},
-			count:          1,
-			groupName:      "my-task",
-			securityGroups: []string{"sg-1", "sg-2"},
-
-			os:   "WINDOWS_SERVER_2019_CORE",
-			arch: "X86_64",
-
-			mockClusterGetter: func(m *mocks.MockClusterGetter) {
-				m.EXPECT().ClusterARN("", "").Return("cluster-1", nil)
-			},
-			mockStarter: func(m *mocks.MockRunner) {
-				m.EXPECT().TaskDefinition("copilot-my-task").Return(&ecs.TaskDefinition{
-					ExecutionRoleArn: aws.String("execution-role"),
-					TaskRoleArn:      aws.String("task-role"),
-					ContainerDefinitions: []*awsecs.ContainerDefinition{
-						{
-							Name:       aws.String("the-one-and-only-one-container"),
-							Image:      aws.String("beautiful-image"),
-							EntryPoint: aws.StringSlice([]string{"enter", "here"}),
-							Command:    aws.StringSlice([]string{"do", "not", "enter", "here"}),
-							Essential:  aws.Bool(true),
-							Environment: []*awsecs.KeyValuePair{
-								{
-									Name:  aws.String("enter"),
-									Value: aws.String("no"),
-								},
-								{
-									Name:  aws.String("kidding"),
-									Value: aws.String("yes"),
-								},
-							},
-							Secrets: []*awsecs.Secret{
-								{
-									Name:      aws.String("truth"),
-									ValueFrom: aws.String("go-ask-the-wise"),
-								},
-							},
-						},
-					},
-				}, nil)
-				m.EXPECT().DescribeTasks("cluster-1", []string{"mockTask1"}).Return([]*ecs.Task{
-					{
-						TaskArn:       aws.String("arn:aws:ecs:us-west-2:123456789:task/4082490ee6c245e09d2145010aa1ba8d"),
-						StoppedReason: aws.String("Task failed to start"),
-						LastStatus:    aws.String("STOPPED"),
-						Containers: []*awsecs.Container{
-							{
-								Name:     aws.String("the-one-and-only-one-container"),
-								ExitCode: aws.Int64(1),
-							},
-						},
-					},
-				}, nil)
-
-			},
-			wantedError: fmt.Errorf("Container the-one-and-only-one-container in task 4082490ee6c245e09d2145010aa1ba8d exited with status code 1"),
-		},
-	}
-
-	for name, tc := range testCases {
-		t.Run(name, func(t *testing.T) {
-			ctrl := gomock.NewController(t)
-			defer ctrl.Finish()
-
-			mockStarter := mocks.NewMockRunner(ctrl)
-			mockClusterGetter := mocks.NewMockClusterGetter(ctrl)
-
-			tc.mockStarter(mockStarter)
-			tc.mockClusterGetter(mockClusterGetter)
-
-			task := &EnvRunner{
-				Count:     tc.count,
-				GroupName: tc.groupName,
-
-				SecurityGroups: tc.securityGroups,
-
-				Starter:       mockStarter,
-				ClusterGetter: mockClusterGetter,
-
-				OS: tc.os,
-			}
-
-			_, err := task.CheckNonZeroExitCode(tc.inTasks)
-			if tc.wantedError != nil {
-				require.EqualError(t, err, tc.wantedError.Error())
-			}
-		})
-	}
-}
