+++ internal/pkg/task/mocks/mock_task.go
@@ -225,48 +225,3 @@ func (mr *MockRunnerMockRecorder) RunTask(input interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RunTask", reflect.TypeOf((*MockRunner)(nil).RunTask), input)
 }
-
-// DescribeTasks mocks base method.
-func (m *MockRunner) DescribeTasks(cluster string, taskARNs []string) ([]*ecs.Task, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "DescribeTasks", cluster, taskARNs)
-	ret0, _ := ret[0].([]*ecs.Task)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// DescribeTasks indicates an expected call of DescribeTasks.
-func (mr *MockRunnerMockRecorder) DescribeTasks(cluster string, taskARNs []string) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DescribeTasks", reflect.TypeOf((*MockRunner)(nil).DescribeTasks), cluster, taskARNs)
-}
-
-// TaskDefinition mocks base method.
-func (m *MockRunner) TaskDefinition(taskDefName string) (*ecs.TaskDefinition, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "TaskDefinition", taskDefName)
-	ret0, _ := ret[0].(*ecs.TaskDefinition)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// TaskDefinition indicates an expected call of TaskDefinition.
-func (mr *MockRunnerMockRecorder) TaskDefinition(taskDefName string) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "TaskDefinition", reflect.TypeOf((*MockRunner)(nil).TaskDefinition), taskDefName)
-}
-
-// CheckNonZeroExitCode mocks base method.
-func (m *MockRunner) CheckNonZeroExitCode(tasks []*ecs.Task) (int, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "CheckNonZeroExitCode", tasks)
-	ret0, _ := ret[0].(int)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// CheckNonZeroExitCode indicates an expected call of CheckNonZeroExitCode.
-func (mr *MockRunnerMockRecorder) CheckNonZeroExitCode(tasks []*ecs.Task) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CheckNonZeroExitCode", reflect.TypeOf((*MockRunner)(nil).CheckNonZeroExitCode), tasks)
-}
